Proper cast

+++ VulkanTest/HelloTriangleApplication.cpp
@@ -282,7 +282,7 @@ VkExtent2D HelloTriangleApplication::ChooseSwapExtent(const VkSurfaceCapabilitie
 	int width, height;
 	glfwGetWindowSize(window_, &width, &height);
 
-	VkExtent2D actualExtent = { width, height };
+	VkExtent2D actualExtent = { static_cast<uint32_t>(width), static_cast<uint32_t>(height) };
 
 	actualExtent.width = std::max(capabilities.minImageExtent.width, std::min(capabilities.maxImageExtent.width, actualExtent.width));
 	actualExtent.height = std::max(capabilities.minImageExtent.height, std::min(capabilities.maxImageExtent.height, actualExtent.height));
